requirements.txt
flask
gunicorn
google-auth-oauthlib
google-api-python-client
requests
Pillow
